requirements.txt
alabaster==0.7.16
altgraph==0.17.4
astroid==3.2.2
Authlib==1.3.1
azure-core==1.34.0
Babel==2.15.0
blinker==1.8.2
Cerberus==1.3.5
certifi==2024.6.2
cffi==1.16.0
click==8.1.7
coverage==7.5.4
cryptography==42.0.8
deepdiff==6.7.1
docutils==0.21.2
et-xmlfile==1.1.0
flake8==7.1.0
Flask==3.0.3
Flask-Cors==4.0.1
google-generativeai==0.5.2
gunicorn==22.0.0
html5lib==1.1
idna==3.7
imagesize==1.4.1
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
lazy-object-proxy==1.9.0
ldap3==2.9.1
MarkupSafe==2.1.5
mccabe==0.7.0
openpyxl==3.1.4
packaging==24.1
Pillow==10.3.0
pluggy==1.5.0
pyasn1==0.6.0
pycodestyle==2.12.0
pycparser==2.22
pycryptodome==3.20.0
pyflakes==3.2.0
Pygments==2.18.0
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
pylint==3.2.3
pymongo==4.7.3
pyOpenSSL==24.1.0
pytest==8.2.2
pytest-cov==5.0.0
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
reportlab==4.2.0
requests==2.32.3
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.3.7
sphinx_rtd_theme==3.0.2
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-httpdomain==1.8.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
urllib3==2.2.2
webencodings==0.5.1
Werkzeug==3.0.3
xhtml2pdf==0.2.16
